Better recognizing when heating cycle reached max
In an uncertain amount of time three subsequent unique temperature values will be recorded, from these we can see if the temperature is rising or dropping. The time the heating should be on to reach a certain temperature will be adjusted every time after the maximum temperature of a cycle has been reached and the temperature starts dropping again.